Return translation key instead of empty string if it doesn't exist.
This is because of localization. Sax2 is translating the descriptions of xkb
properties using it's localization file. English localization file doesn't
contain them, so the original description from xkb should be displayed.
However, because of an obscure way the gettext is used in sax, empty string is
used. Attached patch solves this issue. (bnc#756591)